libbladeRF: set default gain calibration target to 0 on init
Defaulting the gain to 0 is a safer measure to take than grabbing the
current gain. At bladeRF init, the RX gain is set to +60dB which
sets the gain target to +60. If a program depends on the current gain
target while first using the AGC, then the gain will auto adjust to
+60dB + correction when the gain calibration is enabled.

This specifically addresses the case in bladeRF-power where we switch
to manual gain control without setting the gain just yet.